# api/openapi.yaml
openapi: 3.0.3
info:
  title: HeartFolio API
  description: |
    HeartFolio transforms emotional pain and healing processes into verifiable soft-skill assets.

    ## JSON:API

    All endpoints conform to the [JSON:API v1.0](https://jsonapi.org/) specification.

    **Content-Type:** `application/vnd.api+json`

    ### Response envelope

    Every response is wrapped in a top-level `data` object (single resource) or `data` array (collection).

    ```json
    { "data": { "type": "users", "id": "abc123", "attributes": { ... } } }
    ```

    ### Error envelope

    All errors return an `errors` array instead of `data`.

    ```json
    { "errors": [{ "status": "404", "title": "Not Found", "detail": "badge not found" }] }
    ```

    ### Relationships

    Resources that embed related data include a `relationships` key whose value is a full
    resource object (not just a linkage), making it a single-request pattern.

    ### Authentication

    Protected endpoints require `Authorization: Bearer <token>`.
    Obtain a token from `POST /api/v1/users/login` or `POST /api/v1/users/register`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable schemas ──────────────────────────────────────────────────────
  schemas:
...
    ErrorObject:
      type: object
      properties:
        status:
          type: string
          example: "404"
        title:
          type: string
          example: "Not Found"
        detail:
          type: string
          example: "badge not found"
    ErrorDocument:
      type: object
      required: [errors]
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorObject"

    # ── Attribute bags ────────────────────────────────────────────────
...
    BadgeAttributes:
      type: object
      properties:
        name:
          type: string
          example: "Burnout Survivor"
        description:
          type: string
          example: "Awarded for surviving a burnout period"
        category:
          type: string
          example: "resilience"
        max_level:
          type: integer
          format: int32
          example: 5
        criteria:
          type: string
          example: "Complete 3 self-care quests"
        icon:
          type: string
          example: "https://example.com/icons/burnout.png"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserBadgeAttributes:
      type: object
      properties:
        level:
          type: integer
          format: int32
          example: 2
        earned_at:
          type: string
          format: date-time
        last_updated:
          type: string
          format: date-time
    SkillStatsAttributes:
      type: object
      additionalProperties:
        type: integer
        format: int32
      example:
        resilience: 8
        empathy: 5
        self-care: 3
...
    BadgeResource:
      type: object
      properties:
        type:
          type: string
          enum: [badges]
        id:
          type: string
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        attributes:
          $ref: "#/components/schemas/BadgeAttributes"
    UserBadgeResource:
      type: object
      properties:
        type:
          type: string
          enum: [user-badges]
        id:
          type: string
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        attributes:
          $ref: "#/components/schemas/UserBadgeAttributes"
        relationships:
          type: object
          properties:
            badge:
              type: object
              properties:
                data:
                  $ref: "#/components/schemas/BadgeResource"
    SkillStatsResource:
      type: object
      properties:
        type:
          type: string
          enum: [skill-stats]
        id:
          type: string
          description: The user ID
          example: "665f1a2b3c4d5e6f7a8b9c0d"
        attributes:
          $ref: "#/components/schemas/SkillStatsAttributes"
...
    MessageResource:
      type: object
      properties:
        type:
          type: string
          example: "badges"
        id:
          type: string
        attributes:
          type: object
          properties:
            message:
              type: string
              example: "operation successful"

    # ── Top-level documents ────────────────────────────────────────────
...
    BadgeDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/BadgeResource"
    BadgeCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BadgeResource"
    UserBadgeCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserBadgeResource"
    SkillStatsDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/SkillStatsResource"
...
    MessageDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MessageResource"

  # ── Reusable responses ─────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"
          example:
            errors:
              - status: "401"
                title: "Unauthorized"
                detail: "missing authorization header"
    Forbidden:
      description: Insufficient permissions
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"
          example:
            errors:
              - status: "403"
                title: "Forbidden"
                detail: "admin access required"
    NotFound:
      description: Resource not found
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"
          example:
            errors:
              - status: "404"
                title: "Not Found"
                detail: "resource not found"
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/ErrorDocument"

# ── Paths ──────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/badges/{id}:
    get:
      tags: [Badges]
      summary: Get badge by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Badge
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/BadgeDocument"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/badges/users/{userID}/badges:
    get:
      tags: [Badges]
      summary: Get badges for a specific user
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User badge collection with embedded badge details
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserBadgeCollectionDocument"
              example:
                data:
                  - type: user-badges
                    id: "abc"
                    attributes:
                      level: 2
                      earned_at: "2024-01-01T00:00:00Z"
                      last_updated: "2024-01-01T00:00:00Z"
                    relationships:
                      badge:
                        data:
                          type: badges
                          id: "def"
                          attributes:
                            name: "Burnout Survivor"
                            category: "resilience"
                            max_level: 5
  /api/v1/badges/users/{userID}/skill-stats:
    get:
      tags: [Badges]
      summary: Get skill stats for a specific user
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill stats aggregated by badge category
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/SkillStatsDocument"
              example:
                data:
                  type: skill-stats
                  id: "665f1a2b3c4d5e6f7a8b9c0d"
                  attributes:
                    resilience: 8
                    empathy: 5
                    self-care: 3

  # ── Badges (authenticated) ─────────────────────────────────────────────────
  /api/v1/badges/{id}/award:
    post:
      tags: [Badges]
      summary: Award a badge to yourself
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                level:
                  type: integer
                  format: int32
                  minimum: 1
                  default: 1
                  example: 1
      responses:
        "200":
          description: Badge awarded or already at level
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/MessageDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/badges/me/badges:
    get:
      tags: [Badges]
      summary: Get my badges
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user's badge collection
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/UserBadgeCollectionDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/admin/badges/:
    get:
      tags: ["Admin: Badges"]
      summary: List all badges (incl. all categories)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Badge collection
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/BadgeCollectionDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: ["Admin: Badges"]
      summary: Create a badge
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: "Burnout Survivor"
                description:
                  type: string
                category:
                  type: string
                  example: "resilience"
                max_level:
                  type: integer
                  format: int32
                  example: 5
                criteria:
                  type: string
                icon:
                  type: string
      responses:
        "201":
          description: Badge created
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/BadgeDocument"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/badges/{id}:
    get:
      tags: ["Admin: Badges"]
      summary: Get badge by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Badge
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/BadgeDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: ["Admin: Badges"]
      summary: Update a badge
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                category:
                  type: string
                max_level:
                  type: integer
                  format: int32
                criteria:
                  type: string
                icon:
                  type: string
      responses:
        "200":
          description: Badge updated
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/MessageDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: ["Admin: Badges"]
      summary: Delete a badge
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Badge deleted
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/MessageDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/admin/badges/{badgeId}/award:
    post:
      tags: ["Admin: Badges"]
      summary: Award a badge to any user
      security:
        - BearerAuth: []
      parameters:
        - name: badgeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id:
                  type: string
                  example: "665f1a2b3c4d5e6f7a8b9c0d"
                level:
                  type: integer
                  format: int32
                  minimum: 1
                  default: 1
      responses:
        "200":
          description: Badge awarded or already at level
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/MessageDocument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Admin: Quests ──────────────────────────────────────────────────────────
...